requirements.txt
accelerate==0.23.0
bitsandbytes==0.41.2.post2
cachetools==5.3.0
carvekit_colab==4.1.0
dacite==1.8.1
diffusers==0.21.2
ffmpeg_python==0.2.0
imageio>=2.27
lpips==0.1.4
numpy==1.24.2
omegaconf==2.3.0
opencv_python==4.6.0.66
packaging==23.2
Pillow==10.1.0
PyYAML==6.0.1
scikit-image==0.20.0
scipy==1.9.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.34.0
tyro==0.5.7
xformers
tensorboard
